change unsigned int to uint32_t

+++ durationHandler.cpp
@@ -8,22 +8,22 @@ void DurationHandler::setDuration(const Duration& duration) {
 	mMode = Mode::DURATION;
 }
 
-void DurationHandler::setNoOfPeriods(const unsigned int noOfPeriods) {
+void DurationHandler::setNoOfPeriods(const uint32_t noOfPeriods) {
 	mNoOfPeriods = noOfPeriods;
 	mMode = Mode::PERIODS;
 }
 
-void DurationHandler::setNoOfSamples(const unsigned int noOfSamples) {
+void DurationHandler::setNoOfSamples(const uint32_t noOfSamples) {
 	mNoOfSamples = noOfSamples;
 	mMode = Mode::SAMPLES;
 }
 
-uint64_t DurationHandler::calculateNoOfSamples(const uint32_t sampleRate, const uint32_t signalFrequency) const {
+uint32_t DurationHandler::calculateNoOfSamples(const uint32_t sampleRate, const uint32_t signalFrequency) const {
 	if (mMode == Mode::PERIODS) {
-		return static_cast<uint64_t>(mNoOfPeriods) * sampleRate / signalFrequency;
+		return mNoOfPeriods * sampleRate / signalFrequency;
 	}
 	if (mMode == Mode::DURATION) {
-		return static_cast<uint64_t>(mDuration.realSeconds() * sampleRate);
+		return static_cast<uint32_t>(mDuration.realSeconds() * sampleRate);
 	}
 	return mNoOfSamples;
 }
